+++ src/js/view/shared/Skeleton.js
