+++ src/js/view/shared/VideoElement.js
