+++ src/js/view/util/util.js
